implemented structure to make sure that we will have separation of concern
