+++ client/src/index.jsx
@@ -1,7 +1,7 @@
 import React, { useState, createContext } from 'react';
 import ReactDOM from 'react-dom'
 import axios from 'axios'
-import ReviewList from './rating_ and_review/ReviewList.jsx'
+import ReviewIndex from './rating_ and_review/ReviewIndex.jsx'
 
 const UserContext = createContext() // this will help us pass down down without having to chain every single child
 
@@ -24,8 +24,8 @@ const [productId, setProductId] = useState('')
 
   return (
    <div>
-    Webpage
-    <ReviewList></ReviewList>
+    <ReviewIndex />
+
     </div>
   )
 }
