+++ client/src/rating_ and_review/Individual_Review/Overview.jsx
@@ -1,0 +1,14 @@
+import React, { useState } from 'react';
+
+function Overview ({result}) {
+
+
+  return (
+    <div className="eachReview">
+      <p>overview</p>
+
+    </div>
+  )
+}
+
+export default Overview;
