+++ client/src/rating_ and_review/ReviewIndex.jsx
@@ -1,0 +1,16 @@
+import React, { useState, useEffect } from 'react';
+import ReviewList from './Review_List/ReviewList.jsx'
+
+function ReviewIndex(props) {
+
+
+  return (
+    <div>
+      Webpage
+    <ReviewList></ReviewList>
+
+    </div>
+  )
+}
+
+export default ReviewIndex;
