+++ client/src/rating_ and_review/Sorting/Sorting.jsx
@@ -1,0 +1,14 @@
+import React, { useState, useEffect } from 'react';
+//data of reviewList will be passed down here and filter out the reviews here
+
+function Sorting (props) {
+
+
+  return (
+    <div>
+      <h3><code>*num of reviews*</code> reviews, sorted by <code>*dropdown menu for relevance, newest, oldest*</code></h3>
+    </div>
+  )
+}
+
+export default Sorting;
